Reverted cmakelists (Qt version check).
For now, coral-qt is not using qt-4.7.0 because a temporary solution is in use: QMetaType::QVariant constant does not exist in Qt 4.6.2 and its
needed by coral-qt. However, for now, the constant QMetaType::QVarianList is being used as a placeholder temporarily until all build environment is
properly set. The new version will be needed very soon.